third_party/WebKit/LayoutTests/fast/css3-text/css3-text-indent/text-indent-anonymous.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
#container {
  outline: blue solid;
  font-size: 20px;
  line-height: 1;
  text-indent: -3ch;
  width: 5ch;
}
</style>
<p>For the <a href="https://drafts.csswg.org/css-text-3/#text-indent-property">text-indent</a> property,
  the first line of an anonymous block box is only affected if it is the first child of its parent element.</p>
<div id=container>
  <div></div>
  00 00 00
</div>
<script>
test(() => {
  let element = document.getElementById('container');
  assert_equals(element.offsetHeight, 40);
}, "first formatted line of an anonymous block");
</script>
